cleanup stock view
